B2CA-2394 : handle split reception at apdu parsing step, not inside handlers
